+++ base/utils/tests/testthat/test.utils.R
@@ -161,3 +161,135 @@ test_that("mstmipvar works with args specified", {
 #     "Don't know about variable banana in standard_vars in PEcAn.utils"
 #   )
 # })
+
+
+test_that("`left.pad.zeros()` able to add zeros to the left of a number based on `digits`", {
+  expect_equal(left.pad.zeros(123), "00123")
+  expect_equal(left.pad.zeros(42, digits = 3), "042")
+  expect_equal(left.pad.zeros(42, digits = 1), "42")
+})
+
+test_that("`zero.truncate()` able to truncate vector at zero", {
+  input <- c(1, NA, -3, NA, 5)
+  expect_equal(zero.truncate(input), c(1, 0, 0, 0, 5))
+})
+
+test_that("`tabnum()` able to convert positive and negative numbers to `n` significant figures", {
+  
+  # case where n specified
+  x <- c(-2.345, 6.789)
+  result <- tabnum(x, 2)
+  expect_equal(result, c(-2.3, 6.8))
+
+  # case where n is default
+  result <- tabnum(3.5435)
+  expect_equal(result, 3.54)
+})
+
+test_that("`capitalize()` able to capitalize words in a sentence", {
+  # single word
+  expect_equal(capitalize("pecan"), "Pecan")
+
+  # sentence with leading and trailing spaces
+  expect_equal(capitalize("    pecan project    "), "    Pecan Project   ")
+})
+
+test_that("`bibtexify()` able to convert parameters passed to bibtex citation format", {
+  expect_equal(bibtexify("author", "1999", "Here Goes The Title"), "author1999HGTT")
+})
+
+test_that("`rsync()` able to correctly make the command passed to `system` function", {
+  mocked_res <- mockery::mock(0)
+  mockery::stub(rsync, 'system', mocked_res)
+  rsync(args = '-avz', from = 'pecan:test_src', to = 'pecan:test_des')
+  args <- mockery::mock_args(mocked_res)
+  expect_equal(args[[1]][[1]], "rsync -avz pecan:test_src pecan:test_des")
+})
+
+test_that("`ssh()` able to correctly make the command passed to `system` function", {
+  mocked_res <- mockery::mock(0)
+  mockery::stub(ssh, 'system', mocked_res)
+  ssh(host = 'pecan')
+  args <- mockery::mock_args(mocked_res)
+  expect_equal(args[[1]][[1]], "ssh -T pecan \"\" ")
+})
+
+test_that("`temp.settings()` able to create a temporary settings file", {
+  expect_equal(temp.settings('<pecan></pecan>'), '<pecan></pecan>')
+})
+
+test_that("`misc.convert()` able to unit conversions for known and unknown units to the function", {
+
+  # units known to misc.convert
+  expect_equal(misc.convert(1, "kg C m-2 s-1", "umol C m-2 s-1"), 83259094)
+  # units not known to misc.convert
+  expect_equal(misc.convert(10, "kg", "g"), 10000)
+})
+
+test_that("`misc.are.convertible()` able to check if units are convertible by `misc.convert`", {
+  # units known to misc.convert
+  expect_true(misc.are.convertible("kg C m-2 s-1", "umol C m-2 s-1"))
+  # units known but not interconvertible
+  expect_false(misc.are.convertible("kg C m-2 s-1", "Mg ha-1"))
+  # units not known to misc.convert
+  expect_false(misc.are.convertible("kg", "g"))
+})
+
+test_that("`convert.expr()` able to convert expression to variable names", {
+  res <- convert.expr("a+b=c+d")
+  expect_equal(res$variable.drv, "a+b")
+  expect_equal(res$variable.eqn$variables, c("c", "d"))
+  expect_equal(res$variable.eqn$expression, "c+d")
+})
+
+test_that("`paste.stats()` able to print inputs to specific format(for building a Latex Table)", {
+  expect_equal(paste.stats(3.333333, 5.00001, 6.88888, n = 3), "$3.33(5,6.89)$")
+})
+
+test_that("`zero.bounded.density()` returns output containing required components", {
+  res <- zero.bounded.density(c(1, 2, 3))
+  expect_true("x" %in% names(res))
+  expect_true("y" %in% names(res))
+})
+
+test_that("`pdf.stats()` able to calculate mean, variance statistics, and CI from a known distribution", {
+  expect_equal(
+    pdf.stats("beta", 1, 2), 
+    unlist(list(mean = 0.33333333, var = 0.05555556, lcl = 0.01257912, ucl = 0.84188612))
+  )
+})
+
+test_that("`newxtable()` generates correct xtable object", {
+  data <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6))
+  expect_true(grepl("\\hline.*& A & B.*& 1.00 & 4.00.*& 2.00 & 5.00.*& 3.00 & 6.00", newxtable(data)))
+})
+
+test_that("`tryl()` able to check if a function gives an error when called", {
+  # case where function does not give an error
+  expect_true(tryl(1+1))
+
+  # case where function gives an error
+  expect_false(tryl(log("a")))
+})
+
+test_that("`download_file()` able to correctly construct the inputs command to system function", {
+  mocked_res <- mockery::mock(0)
+  mockery::stub(download_file, 'system', mocked_res)
+  download_file("ftp://testpecan.com", "test", "ncftpget")
+  args <- mockery::mock_args(mocked_res)
+  expect_equal(args[[1]][[1]], "ncftpget -c ftp://testpecan.com > test")
+})
+
+test_that("`retry.func()` able to retry a function before returning an error", {
+  defaultW <- getOption("warn") 
+  options(warn = -1)
+  on.exit(options(warn = defaultW))
+  expect_error(
+    retry.func(ncdf4::nc_open("http://pecan"), maxErrors = 2, sleep = 2),
+    "retry: too many retries"
+  )
+
+  # case where function does not give an error
+  expect_equal(retry.func(1+1, maxErrors = 2, sleep = 2), 2)
+})
+